improves input validation and limits framed msg
